fix expiration int bug

+++ src/main/java/com/vechain/thorclient/core/model/blockchain/Transaction.java
@@ -9,7 +9,7 @@ public class Transaction implements Serializable {
     private int   size;
     private int   chainTag;
     private String  blockRef; //8 bytes
-    private int expiration;
+    private long expiration;
     private ArrayList<Clause> clauses;
 
     private int gasPriceCoef;
@@ -58,11 +58,11 @@ public class Transaction implements Serializable {
         this.blockRef = blockRef;
     }
 
-    public int getExpiration() {
+    public long getExpiration() {
         return expiration;
     }
 
-    public void setExpiration(int expiration) {
+    public void setExpiration(long expiration) {
         this.expiration = expiration;
     }
